Abstract away location and format of file-based semaphores
The code originally used strings of the form "/tmp/..." to denote that
semaphore files should go in `/tmp`. Now we call
 `Sl_Utils::slip_semaphore_file` to build up the filename programatically
and place them in SLIP_TMP_DIR (currently `/tmp/slip_semaphores/`.

The SLIP_TMP_DIR is created using `make_path` in a BEGIN block on compile
so it should alwys be there.