Put expected username running tomcats into configuration 'tomcats_run_as_user'
The procedure which determines whether tomcat is running or not calls
out to `pgrep` and restricts it by a userid. This was hardcoded (to
'tomcat'), and the last round of upgrades changed it (to 'tomcatrhel5').

This adds a configuration key, 'tomcats_run_as_user', to runs 10,11, and 60
and uses it in the is_tomcat_running() method.