ETT-114: Remove all traces of publish on demand
We want to test access as if we are located in the US, but that triggers
a lookup for the pod table, which doesn't exist in dbimage.

Given that we have decomissioned this service, I decided it was time to
remove the code that queries it and that displays the links.