First try at playwright tests for authenticated access
This is rather janky, but suffices for a first attempt to demonstrate.
First, we ets playwright tests with "auth" in the name not to run with
npm run test (using '--grep-invert').

To test:

* run switch_auth.ssh, pick ssduser
* manually reload apache-test config: docker compose exec apache-test
  kill -USR1 1
* run playwright tests with:
    docker compose run playwright npx playwright test ssduser --trace on
* should pass

Next step is automating running the authed tests in github actions
(switch_auth.sh will need some updates) and maybe seeing if there's a
better way to filter tests.