Update pdfbox for the 'stamper'
This "stamper" is only used when there is an existing PDF in the
repository we want to serve (right now only knowledge unlatched), but I
didn't realize that when I tried to build/update it. This functionality
seems broken in production right now, and I don't know that it's worth
debugging right now. There are probably other dependencies we
could/should update here too.