ETT-217: hide limited view search results when limit parameter is ft
In some cases we may output items that are actually limited view even
when we have the fulltext facet selected. In particular, for resource
sharing because of the way the holdings filter works we may select items
that are withdrawn, lost, or missing, which resource sharing users don't
actually get access to.